migrate schema utilities to a different location
