+++ tests/test_scripts.py
@@ -63,7 +63,7 @@ def test_import_prepHistFactory(tmpdir, script_runner):
 
     parsed_xml = json.loads(temp.read())
     spec = {'channels': parsed_xml['channels']}
-    pyhf.utils.validate(spec, 'model.json')
+    pyhf.schema.validate(spec, 'model.json')
 
 
 def test_import_prepHistFactory_withProgress(tmpdir, script_runner):
